# Drop BsonClassMap AutoMap