--- a/src/MongoFramework/Infrastructure/Mapping/DefaultMappingPack.cs
+++ b/src/MongoFramework/Infrastructure/Mapping/DefaultMappingPack.cs
@@ -13,13 +13,12 @@ namespace MongoFramework.Infrastructure.Mapping
 			{
 				new CollectionNameProcessor(),
 				new HierarchyProcessor(),
+				new PropertyMappingProcessor(),
 				new EntityIdProcessor(),
-				new MappedPropertiesProcessor(),
 				new NestedTypeProcessor(),
 				new ExtraElementsProcessor(),
 				new TypeDiscoveryProcessor(),
 				new BsonKnownTypesProcessor(),
-				new ClassMapPropertiesProcessor(),
 				new IndexProcessor(),
 				new EntityRelationshipProcessor()
 			};