--- a/src/MongoFramework/Infrastructure/Mapping/Processors/EntityIdProcessor.cs
+++ b/src/MongoFramework/Infrastructure/Mapping/Processors/EntityIdProcessor.cs
@@ -15,21 +15,21 @@ namespace MongoFramework.Infrastructure.Mapping.Processors
 			var entityType = definition.EntityType;
 
 			//Find the first property with the "Key" attribute to use as the Id
-			var properties = entityType.GetProperties(BindingFlags.Public | BindingFlags.Instance | BindingFlags.DeclaredOnly);
-			var idProperty = properties.Where(p => p.GetCustomAttribute<KeyAttribute>() != null).FirstOrDefault();
+			var properties = definition.Properties;
+			var idProperty = properties.Where(p => p.PropertyInfo.GetCustomAttribute<KeyAttribute>() != null).FirstOrDefault();
 			if (idProperty != null)
 			{
-				classMap.MapIdMember(idProperty);
+				classMap.MapIdMember(idProperty.PropertyInfo);
 			}
 			else
 			{
 				idProperty = properties
-					.Where(p => p.Name.Equals("id", StringComparison.InvariantCultureIgnoreCase))
+					.Where(p => p.ElementName.Equals("id", StringComparison.InvariantCultureIgnoreCase))
 					.FirstOrDefault();
 
 				if (idProperty != null)
 				{
-					classMap.MapIdMember(idProperty);
+					classMap.MapIdMember(idProperty.PropertyInfo);
 				}
 			}