--- a/src/MongoFramework/Infrastructure/Mapping/Processors/ExtraElementsProcessor.cs
+++ b/src/MongoFramework/Infrastructure/Mapping/Processors/ExtraElementsProcessor.cs
@@ -23,17 +23,17 @@ namespace MongoFramework.Infrastructure.Mapping.Processors
 			else
 			{
 				//If any of the Entity's properties have the "ExtraElementsAttribute", assign that against the BsonClassMap
-				var extraElementsProperty = entityType.GetProperties(BindingFlags.Public | BindingFlags.Instance | BindingFlags.DeclaredOnly)
+				var extraElementsProperty = definition.Properties
 					.Select(p => new
 					{
-						PropertyInfo = p,
-						ExtraElementsAttribute = p.GetCustomAttribute<ExtraElementsAttribute>()
+						Property = p,
+						ExtraElementsAttribute = p.PropertyInfo.GetCustomAttribute<ExtraElementsAttribute>()
 					}).Where(p => p.ExtraElementsAttribute != null).FirstOrDefault();
 
-				if (extraElementsProperty != null && typeof(IDictionary<string, object>).IsAssignableFrom(extraElementsProperty.PropertyInfo.PropertyType))
+				if (extraElementsProperty != null && typeof(IDictionary<string, object>).IsAssignableFrom(extraElementsProperty.Property.PropertyType))
 				{
 					var memberMap = classMap.DeclaredMemberMaps
-						.Where(m => m.MemberInfo == extraElementsProperty.PropertyInfo)
+						.Where(m => m.ElementName == extraElementsProperty.Property.ElementName)
 						.FirstOrDefault();
 					classMap.SetExtraElementsMember(memberMap);
 				}