--- a/tests/MongoFramework.Benchmarks/MongoFramework.Benchmarks.csproj
+++ b/tests/MongoFramework.Benchmarks/MongoFramework.Benchmarks.csproj
@@ -14,8 +14,4 @@
     <ProjectReference Include="..\..\src\MongoFramework\MongoFramework.csproj" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="docs\" />
-  </ItemGroup>
-
 </Project>