… src/MongoFramework/Infrastructure/Mapping/Processors/ClassMapPropertiesProcessor.cs
… src/MongoFramework/Infrastructure/Mapping/Processors/MappedPropertiesProcessor.cs
… tests/MongoFramework.Tests/Infrastructure/Mapping/Processors/MappedPropertiesProcessorTests.cs