Closes #106 

**IndexModelBuilderBenchmark**

*Before*

|      Method |     Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------ |---------:|----------:|----------:|-------:|------:|------:|----------:|
|   FlatModel | 33.13 us | 0.2674 us | 0.2502 us | 3.7842 |     - |     - |  11.59 KB |
| NestedModel | 37.18 us | 0.4619 us | 0.4094 us | 4.9438 |     - |     - |  15.15 KB |

*After*

|      Method |     Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------ |---------:|----------:|----------:|-------:|------:|------:|----------:|
|   FlatModel | 32.03 us | 0.3295 us | 0.2752 us | 3.6621 |     - |     - |  11.25 KB |
| NestedModel | 35.67 us | 0.3898 us | 0.3646 us | 4.6997 |     - |     - |   14.5 KB |

**EntityMappingBenchmark**

*Before*

|                Method |        Mean |       Error |       StdDev |   Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------- |------------:|------------:|-------------:|--------:|------:|------:|----------:|
|         ResetOverhead |    973.1 ns |    10.02 ns |     9.369 ns |  0.2346 |     - |     - |     737 B |
|       ClassMapAutoMap | 26,108.3 ns |   129.99 ns |   108.549 ns |  2.6550 |     - |     - |    8373 B |
| EntityMappingRegister | 98,133.5 ns | 1,355.18 ns | 1,267.632 ns | 12.6953 |     - |     - |   40571 B |

*After*

|                Method |        Mean |      Error |     StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------- |------------:|-----------:|-----------:|-------:|------:|------:|----------:|
|         ResetOverhead |    991.8 ns |   8.703 ns |   8.141 ns | 0.2346 |     - |     - |     737 B |
|       ClassMapAutoMap | 26,475.4 ns | 284.697 ns | 266.306 ns | 2.6550 |     - |     - |    8373 B |
| EntityMappingRegister | 46,178.4 ns | 303.181 ns | 283.596 ns | 9.1553 |     - |     - |   28906 B |